requirements.txt
Flask==2.3.3
gunicorn==20.1.0
Werkzeug==2.3.7
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
flask-smorest==0.42.1
marshmallow==3.19.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.6.0
passlib==1.7.4
python-dateutil==2.8.2